+++ src/hardfile_host.cpp
@@ -102,9 +102,9 @@ static void rdbdump(FILE* h, uae_u64 offset, uae_u8* buf, int blocksize) {
     for (i = 0; i <= blocks; i++) {
         if (fseeko64(h, offset, SEEK_SET) != 0)
             break;
-        int outlen = fread(buf, 1, blocksize, h);
-        if (outlen != blocksize) {
-            write_log("rdbdump: warning: read %d bytes (not blocksize %d)\n", outlen, blocksize);
+        size_t outlen = fread(buf, 1, blocksize, h);
+        if (outlen != static_cast<size_t>(blocksize)) {
+            write_log("rdbdump: warning: read %zu bytes (not blocksize %d)\n", outlen, blocksize);
         }
         fwrite(buf, 1, blocksize, f);
         offset += blocksize;
@@ -128,7 +128,7 @@ static int ismounted(FILE* f) {
 static int safetycheck(FILE* h, const char* name, uae_u64 offset, uae_u8* buf, int blocksize) {
     (void)name;
     int i, j, blocks = 63, empty = 1;
-    long outlen;
+    size_t outlen;
 
     for (j = 0; j < blocks; j++) {
         if (fseeko64(h, offset, SEEK_SET) != 0) {
@@ -137,7 +137,7 @@ static int safetycheck(FILE* h, const char* name, uae_u64 offset, uae_u8* buf, i
         }
         memset(buf, 0xaa, blocksize);
         outlen = fread(buf, 1, blocksize, h);
-        if (outlen != blocksize) {
+        if (outlen != static_cast<size_t>(blocksize)) {
             write_log("hd ignored, read error %d!\n", errno);
             return 2;
         }
@@ -293,7 +293,7 @@ int hdf_open_target(struct hardfiledata* hfd, const char* pname) {
         if (h == INVALID_HANDLE_VALUE)
             goto end;
         hfd->handle->h = h;
-        i = _tcslen(name) - 1;
+        i = static_cast<int>(_tcslen(name)) - 1;
         while (i >= 0) {
             if ((i > 0 && (name[i - 1] == '/' || name[i - 1] == '\\')) || i == 0) {
                 _tcscpy(hfd->vendor_id, "UAE");
@@ -306,7 +306,7 @@ int hdf_open_target(struct hardfiledata* hfd, const char* pname) {
         if (h != INVALID_HANDLE_VALUE) {
             // determine size of hdf file
             int ret;
-            off_t low = -1;
+            int64_t low = -1;
 
 #if defined(MACOSX) || defined(OPENBSD)
             // check type of file
@@ -535,7 +535,7 @@ void hfd_flush_cache (struct hardfiledata *hfd, int now)
 #endif
 
 static int hdf_read_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int len) {
-    long outlen = 0;
+    size_t outlen = 0;
     int coffset;
 
     if (offset == 0)
@@ -555,7 +555,7 @@ static int hdf_read_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, in
     else if (hfd->handle_valid == HDF_HANDLE_ZFILE)
         outlen = zfile_fread(hfd->cache, 1, CACHE_SIZE, hfd->handle->zf);
     hfd->cache_valid = 0;
-    if (outlen != CACHE_SIZE)
+    if (outlen != static_cast<size_t>(CACHE_SIZE))
         return 0;
     hfd->cache_valid = 1;
     coffset = isincache(hfd, offset, len);
@@ -578,7 +578,7 @@ int hdf_read_target(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int 
         return 0;
     while (len > 0) {
         int maxlen;
-        int ret = 0;
+        size_t ret = 0;
         if (hfd->physsize < CACHE_SIZE) {
             hfd->cache_valid = 0;
             hdf_seek(hfd, offset);
@@ -592,10 +592,10 @@ int hdf_read_target(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int 
             maxlen = len;
         } else {
             maxlen = len > CACHE_SIZE ? CACHE_SIZE : len;
-            ret = hdf_read_2(hfd, p, offset, maxlen);
+            ret = static_cast<size_t>(hdf_read_2(hfd, p, offset, maxlen));
         }
-        got += ret;
-        if (ret != maxlen)
+        got += static_cast<int>(ret);
+        if (ret != static_cast<size_t>(maxlen))
             return got;
         offset += maxlen;
         p += maxlen;
@@ -605,7 +605,7 @@ int hdf_read_target(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int 
 }
 
 static int hdf_write_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int len) {
-    int outlen = 0;
+    size_t outlen = 0;
 
     if (hfd->ci.readonly) {
         if (g_debug) {
@@ -627,11 +627,11 @@ static int hdf_write_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, i
         outlen = fwrite(hfd->cache, 1, len, hfd->handle->h);
         // fflush(hfd->handle->h);
         if (g_debug) {
-            write_log("wrote %u bytes (wanted %d) at offset %llx\n", outlen, len, (unsigned long long)offset);
+            write_log("wrote %zu bytes (wanted %d) at offset %llx\n", outlen, len, (unsigned long long)offset);
         }
         const TCHAR* name = hfd->emptyname == NULL ? _T("<unknown>") : hfd->emptyname;
         if (offset == 0) {
-            long outlen2;
+            size_t outlen2;
             (void)outlen2; // Suppress set-but-not-used warning
             uae_u8* tmp;
             int tmplen = 512;
@@ -640,7 +640,7 @@ static int hdf_write_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, i
                 memset(tmp, 0xa1, tmplen);
                 hdf_seek(hfd, offset);
                 outlen2 = fread(tmp, 1, tmplen, hfd->handle->h);
-                if (memcmp(hfd->cache, tmp, tmplen) != 0 || outlen != len)
+                if (memcmp(hfd->cache, tmp, tmplen) != 0 || outlen != static_cast<size_t>(len))
                     gui_message(_T("\"%s\"\n\nblock zero write failed!"), name);
                 xfree(tmp);
             }
@@ -648,7 +648,7 @@ static int hdf_write_2(struct hardfiledata* hfd, void* buffer, uae_u64 offset, i
     } else if (hfd->handle_valid == HDF_HANDLE_ZFILE) {
         outlen = zfile_fwrite(hfd->cache, 1, len, hfd->handle->zf);
     }
-    return outlen;
+    return static_cast<int>(outlen);
 }
 int hdf_write_target(struct hardfiledata* hfd, void* buffer, uae_u64 offset, int len) {
     int got = 0;
